// packages/Common/Types/Service/TechStack.ts
export enum TechStack {
  NodeJS = "NodeJS",
  React = "React",
  Python = "Python",
  Ruby = "Ruby",
  Go = "Go",
  Java = "Java",
  PHP = "PHP",
  CSharp = "C#",
  CPlusPlus = "C++",
  Rust = "Rust",
  Swift = "Swift",
  Kotlin = "Kotlin",
  TypeScript = "TypeScript",
  JavaScript = "JavaScript",
  Shell = "Shell",
  Markdown = "Markdown",
  Other = "Other",
}

export default TechStack;